Add files via upload
---EXPLANATION OF PROGRAM---
The initial purpose of the single_wave_phase_distance_final.mlx program is to be added to a SFCW radar system with an object in scene. It takes recieved radar waves across a set bandwidth (in this case, 71-79GHz), and is able to, using amplitude and phase data, find the distance to the target to an exact degree. Using just amplitude, the resolution would be 2 cm, yet with this program, the resolution would be to whatever decimal MatLab rounds to, which is 1.0 * 10^-14 cm. These distances would be generated in one dimension, however to use this program before an azimuthal FFT, this could be changed to facilitate 2 or 3 dimensional measurements.

Since this program was developed seperately from a SFCW radar, and meant to be implemented later, the radar waves are created digitally from a wave equation and a 'target distance' variable. These can easily be removed to connect to a real SFCW radar, as as the input waves would be digital.

---HOW TO READ RESULTS---
Input any distance you want into the 'target_distance' variable that is below the maximum unambiguous range. The program will then output the found distance from phase in the 'final distance' variable. Note that at high distances, specifically less than half a wavelength from the unambiguous range, this breaks down and will output half the actual distance. This is a quick fix, but I will need to run through the program with large numbers to do.